+++ resolwe/test/__init__.py
@@ -12,11 +12,12 @@ Resolwe Test Framework
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from resolwe.test.testcases import ElasticSearchTestCase, ProcessTestCase, ResolweAPITestCase, TestCase
+from resolwe.test.testcases import (
+    ElasticSearchTestCase, ProcessTestCase, ResolweAPITestCase, TestCase, TransactionElasticSearchTestCase,
+)
 from resolwe.test.utils import check_docker, check_installed, with_docker_executor
 
 __all__ = (
-    'TestCase', 'ProcessTestCase', 'ResolweAPITestCase',
-    'ElasticSearchTestCase', 'check_installed', 'check_docker',
-    'with_docker_executor',
+    'TestCase', 'ProcessTestCase', 'ResolweAPITestCase', 'ElasticSearchTestCase',
+    'TransactionElasticSearchTestCase', 'check_installed', 'check_docker', 'with_docker_executor',
 )